requirements.txt
ply
pytest
pymongo
psycopg2-binary
psycopg
flask
